refactor(data): repackage data
