+++ app/src/main/java/com/blueprint/database/AppDatabase.kt
@@ -2,10 +2,10 @@ package com.blueprint.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.blueprint.featureA.data.local_model.LocalFeatureA
-import com.blueprint.featureA.service.FeatureALocalService
+import com.blueprint.featureA.data.model.local.FeatureAEntity
+import com.blueprint.featureA.data.sources.local.FeatureADao
 
-@Database(entities = [LocalFeatureA::class], version = 1)
+@Database(entities = [FeatureAEntity::class], version = 1)
 abstract class AppDatabase : RoomDatabase() {
-    abstract fun featureADao(): FeatureALocalService
+    abstract fun featureADao(): FeatureADao
 }
