+++ app/src/main/java/com/blueprint/featureA/data/model/local/FeatureAEntity.kt
@@ -1,10 +1,10 @@
-package com.blueprint.featureA.data.local_model
+package com.blueprint.featureA.data.model.local
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "feature_a")
-data class LocalFeatureA(
+data class FeatureAEntity(
     @PrimaryKey(autoGenerate = true) val id: Int,
     val string: String,
 )
