+++ app/src/main/java/com/blueprint/featureA/data/model/remote/NetworkFeatureA.kt
@@ -1,0 +1,5 @@
+package com.blueprint.featureA.data.model.remote
+
+data class NetworkFeatureA(
+    val string: String
+)
