--- app/src/main/java/com/blueprint/featureA/data/remote_model/RemoteFeatureA.kt
@@ -1,5 +1,0 @@
-package com.blueprint.featureA.data.remote_model
-
-data class RemoteFeatureA(
-    val string: String
-)
