+++ app/src/main/java/com/blueprint/featureA/mapper/DefaultFeatureARemoteRemoteMapper.kt
@@ -1,12 +1,12 @@
 package com.blueprint.featureA.mapper
 
-import com.blueprint.featureA.data.remote_model.RemoteFeatureA
+import com.blueprint.featureA.data.model.remote.NetworkFeatureA
 import com.blueprint.featureA.domain.model.FeatureA
 
 class DefaultFeatureARemoteRemoteMapper : FeatureARemoteMapper {
-    override fun map(featuresA: List<RemoteFeatureA>): List<FeatureA> =
+    override fun map(featuresA: List<NetworkFeatureA>): List<FeatureA> =
         featuresA.map { map(it) }
 
-    override fun map(featureA: RemoteFeatureA): FeatureA =
+    override fun map(featureA: NetworkFeatureA): FeatureA =
         FeatureA(featureA.string)
 }
