+++ app/src/main/java/com/blueprint/featureA/mapper/FeatureALocalMapper.kt
@@ -1,10 +1,10 @@
 package com.blueprint.featureA.mapper
 
-import com.blueprint.featureA.data.local_model.LocalFeatureA
+import com.blueprint.featureA.data.model.local.FeatureAEntity
 import com.blueprint.featureA.domain.model.FeatureA
 
 interface FeatureALocalMapper {
 
-    fun map(featuresA: List<LocalFeatureA>): List<FeatureA>
-    fun map(featureA: LocalFeatureA): FeatureA
+    fun map(featuresA: List<FeatureAEntity>): List<FeatureA>
+    fun map(featureA: FeatureAEntity): FeatureA
 }
