+++ app/src/main/java/com/blueprint/featureA/mapper/FeatureARemoteMapper.kt
@@ -1,11 +1,11 @@
 package com.blueprint.featureA.mapper
 
-import com.blueprint.featureA.data.remote_model.RemoteFeatureA
+import com.blueprint.featureA.data.model.remote.NetworkFeatureA
 import com.blueprint.featureA.domain.model.FeatureA
 
 interface FeatureARemoteMapper {
 
-    fun map(featuresA: List<RemoteFeatureA>): List<FeatureA>
-    fun map(featureA: RemoteFeatureA): FeatureA
+    fun map(input: List<NetworkFeatureA>): List<FeatureA>
+    fun map(input: NetworkFeatureA): FeatureA
 
 }
