+++ app/src/main/java/com/blueprint/featureA/repository/DefaultFeatureALocalRepository.kt
@@ -1,21 +1,21 @@
 package com.blueprint.featureA.repository
 
-import com.blueprint.featureA.data.local_model.LocalFeatureA
-import com.blueprint.featureA.data_sources.local.FeatureADao
+import com.blueprint.featureA.data.model.local.FeatureAEntity
+import com.blueprint.featureA.data.sources.local.FeatureADao
 import javax.inject.Inject
 
 class DefaultFeatureALocalRepository @Inject constructor(
     private val featureADao: FeatureADao
 ) : FeatureALocalRepository{
-    override suspend fun insertFeatureA(localFeatureA: LocalFeatureA) {
+    override suspend fun insertFeatureA(featureAEntity: FeatureAEntity) {
         featureADao.insertFeatureA(featureAEntity)
     }
 
-    override suspend fun updateFeatureA(localFeatureA: LocalFeatureA) {
+    override suspend fun updateFeatureA(featureAEntity: FeatureAEntity) {
         featureADao.updateFeatureA(featureAEntity)
     }
 
-    override suspend fun getFeatureAById(id: Int): LocalFeatureA? {
+    override suspend fun getFeatureAById(id: Int): FeatureAEntity? {
         return featureADao.getFeatureAById(id)
     }
