+++ app/src/main/java/com/blueprint/featureA/repository/DefaultFeatureARemoteRepository.kt
@@ -1,6 +1,6 @@
 package com.blueprint.featureA.repository
 
-import com.blueprint.featureA.data_sources.remote.FeatureAService
+import com.blueprint.featureA.data.sources.remote.FeatureAService
 import com.blueprint.featureA.domain.model.FeatureA
 import com.blueprint.featureA.mapper.FeatureARemoteMapper
 import javax.inject.Inject
