+++ app/src/main/java/com/blueprint/featureA/repository/FeatureALocalRepository.kt
@@ -1,10 +1,10 @@
 package com.blueprint.featureA.repository
 
-import com.blueprint.featureA.data.local_model.LocalFeatureA
+import com.blueprint.featureA.data.model.local.FeatureAEntity
 
 interface FeatureALocalRepository {
-    suspend fun insertFeatureA(localFeatureA: LocalFeatureA)
-    suspend fun updateFeatureA(localFeatureA: LocalFeatureA)
-    suspend fun getFeatureAById(id: Int): LocalFeatureA?
+    suspend fun insertFeatureA(featureAEntity: FeatureAEntity)
+    suspend fun updateFeatureA(featureAEntity: FeatureAEntity)
+    suspend fun getFeatureAById(id: Int): FeatureAEntity?
     suspend fun deleteFeatureAById(id: Int)
 }
